+++ src/test/java/org/oba/jedis/extra/utils/cache/MockOfJedisForSimpleCacheTest.java
@@ -1,4 +1,4 @@
-package org.oba.jedis.extra.utils.simple;
+package org.oba.jedis.extra.utils.cache;
 
 import org.junit.After;
 import org.junit.Before;
@@ -17,12 +17,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({Transaction.class, TransactionBase.class })
