+++ src/test/java/org/oba/jedis/extra/utils/cache/SimpleCacheWriterTest.java
@@ -1,54 +1,43 @@
-package org.oba.jedis.extra.utils.simple.functional;
+package org.oba.jedis.extra.utils.cache;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-import org.oba.jedis.extra.utils.cache.CacheWriter;
-import org.oba.jedis.extra.utils.cache.SimpleCache;
-import org.oba.jedis.extra.utils.test.JedisTestFactory;
 import org.oba.jedis.extra.utils.utils.SimpleEntry;
-import redis.clients.jedis.JedisPool;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
+import redis.clients.jedis.Transaction;
+import redis.clients.jedis.TransactionBase;
 
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
+import static org.oba.jedis.extra.utils.cache.MockOfJedisForSimpleCache.unitTestEnabledForSimpleCache;
 
-@RunWith(MockitoJUnitRunner.Silent.class)
-public class FunctionalSimpleCacheWriterTest {
+//@RunWith(MockitoJUnitRunner.Silent.class)
+@RunWith(PowerMockRunner.class)
+@PrepareForTest({Transaction.class, TransactionBase.class })
+public class SimpleCacheWriterTest {
 
-
-
-
-    private final JedisTestFactory jtfTest = JedisTestFactory.get();
+    private MockOfJedisForSimpleCache mockOfJedisForSimpleCache;
 
     private final TestingCacheWriter testingCacheWriter = new TestingCacheWriter();
 
-    private JedisPool jedisPool;
-
     @Before
     public void setup() {
-        org.junit.Assume.assumeTrue(jtfTest.functionalTestEnabled());
-        if (!jtfTest.functionalTestEnabled()) return;
-        jedisPool = jtfTest.createJedisPool();
+        org.junit.Assume.assumeTrue(unitTestEnabledForSimpleCache());
+        if (!unitTestEnabledForSimpleCache()) return;
+        mockOfJedisForSimpleCache = new MockOfJedisForSimpleCache();
     }
 
     @After
     public void tearDown() {
-        if (jedisPool != null) {
-            jedisPool.close();
+        if (mockOfJedisForSimpleCache != null) {
+            mockOfJedisForSimpleCache.clearData();
         }
     }
 
@@ -58,11 +47,9 @@ public class FunctionalSimpleCacheWriterTest {
 
     SimpleCache createNewCache(CacheWriter cacheWriter) {
         String name = "cache:" + this.getClass().getName() + ":" + System.currentTimeMillis();
-        SimpleCache simpleCache = new SimpleCache(jedisPool, name, 3_600_000).
-                withCacheWriter(testingCacheWriter);
-        return simpleCache;
+        return new SimpleCache(mockOfJedisForSimpleCache.getJedisPool(), name, 3_600_000)
+                .withCacheWriter(cacheWriter);
     }
-
 
     @Test
     public void putTest() {
@@ -96,6 +83,24 @@ public class FunctionalSimpleCacheWriterTest {
     @Test
     public void putAllTest() {
         SimpleCache simpleCache = createNewCache();
+        Map<String, String> data = new HashMap<>();
+        data.put("a", "A1");
+        data.put("b", "B1");
+        data.put("c", "C1");
+        simpleCache.putAll(data);
+        assertEquals("A1", simpleCache.get("a"));
+        assertEquals("A1", testingCacheWriter.get("a"));
+        assertEquals("B1", simpleCache.get("b"));
+        assertEquals("B1", testingCacheWriter.get("b"));
+        assertEquals("C1", simpleCache.get("c"));
+        assertEquals("C1", testingCacheWriter.get("c"));
+        assertEquals(3, testingCacheWriter.countDataInserted());
+        assertEquals(0, testingCacheWriter.countDataDeleted());
+    }
+
+    @Test
+    public void defaultPutAllTest() {
+        SimpleCache simpleCache = createNewCache(new TestingDefaultsCacheWriter(testingCacheWriter));
         Map<String, String> data = new HashMap<>();
         data.put("a", "A1");
         data.put("b", "B1");
@@ -285,15 +290,15 @@ public class FunctionalSimpleCacheWriterTest {
 
     @Test(expected = IllegalStateException.class)
     public void removeAllWithErrorTest() {
-        SimpleCache simpleCache = createNewCache();
+        SimpleCache SimpleCache = createNewCache();
         testingCacheWriter.doNextError();
         Map<String, String> data = new HashMap<>();
         data.put("a", "A1");
         data.put("b", "B1");
         data.put("c", "C1");
         data.put("d", "D1");
-        simpleCache.putAll(data);
-        simpleCache.removeAll(new HashSet<>(Arrays.asList("a","b","c")));
+        SimpleCache.putAll(data);
+        SimpleCache.removeAll(new HashSet<>(Arrays.asList("a","b","c")));
     }
 
     @Test
@@ -319,7 +324,7 @@ public class FunctionalSimpleCacheWriterTest {
     }
 
     @Test
-    public void clearlTest() {
+    public void clearTest() {
         SimpleCache simpleCache = createNewCache();
         Map<String, String> data = new HashMap<>();
         data.put("a", "A1");
@@ -345,11 +350,11 @@ public class FunctionalSimpleCacheWriterTest {
         TestingCacheWriter localCacheWriter = new TestingCacheWriter();
         localCacheWriter.write(new SimpleEntry("a","A1"));
         localCacheWriter.write("b","B1");
-        Map<String, String> datas = new HashMap<>();
-        datas.put("c","C1");
-        datas.put("d","D1");
-        datas.put("e","E1");
-        localCacheWriter.writeAll(datas);
+        Map<String, String> data = new HashMap<>();
+        data.put("c","C1");
+        data.put("d","D1");
+        data.put("e","E1");
+        localCacheWriter.writeAll(data);
         assertEquals("A1", localCacheWriter.get("a"));
         assertEquals("B1", localCacheWriter.get("b"));
         assertEquals("C1", localCacheWriter.get("c"));
@@ -368,12 +373,6 @@ public class FunctionalSimpleCacheWriterTest {
         assertNull(localCacheWriter.get("d"));
         assertNull(localCacheWriter.get("e"));
     }
-
-    /*
-    removeAll
-            removeAll
-*/
-
 
     private static class TestingCacheWriter implements CacheWriter {
 
@@ -406,6 +405,7 @@ public class FunctionalSimpleCacheWriterTest {
         public boolean contains(String key) {
             return internalData.containsKey(key);
         }
+
 
         @Override
         public void write(String key, String value) {
@@ -446,7 +446,7 @@ public class FunctionalSimpleCacheWriterTest {
 
     private static class TestingDefaultsCacheWriter implements CacheWriter {
 
-        private TestingCacheWriter testingCacheWriter;
+        private final TestingCacheWriter testingCacheWriter;
 
         public TestingDefaultsCacheWriter(TestingCacheWriter testingCacheWriter) {
             this.testingCacheWriter = testingCacheWriter;
